Produce an indexed library of multi-extension fits file for PhoSim images
PhoSim delivers Psfs as grids of stars, one for each Rnn_Smm.fits.gz.
This routine creates mini-libraries named Rnn.psfs.fits, one for
each raft, plus an index named .index.  The suffix of these files is the
same as the suffix of the PhoSim files fits.gz files from which the
library is created.

A sample of running this file is in bin/makelibrary