Changes to allow the processShearTest to work with new libraries
Libraries now calls psf/psf_library_nn.fits, and locate in the
directory which contains the GalSim output.  This is the "input"
parameter for the processShearTest task.
The psf_library and psf_index values in the GalSim catalog
identify the library and hdu used for the Psf of each source.