These changes reorganize the examples to match the run_params scheme
run_params is a file which must be in the great3sims run directory.
It is used by all of the scripts, and communicates the params between.
The base files no longer need to be checked in - generated by runGal
A run_params file which can be used with base is stored in examples.
It is copied to the base directory when it is made.
The psfs are created and stored in a separate directory, sample_psfs
A link to them is placed in base/control/ground/constant before runGreat.
I have checked in the psfs index and library files, but they can
be created with makelibrary