Changes to this routine were made to alter the error estimation to match Numerical recipes 15.2
The big change to the code was to no longer use the covariance matrix from the fitting routine,
which does not correctly reflect the errors in the fit based on errors in the shear averages as
describe below.  Instead, use the error estimation in Numerical recipes 15.2 for calculating
the slope and intercept error for a first order fit.

Also, remove the code to calculate average deviations e1-g1 and e2-g2 for individual subfields
as a way to estimate the stdev of those measurements for each subfield.  The code now just
estimates the mean value and stdev of e1 and e2.