Modify looping of base directories to allow for multiple output runs on the same data
A test is a named output directory on a give set of simulation images.
The src.fits files from the measurement and the analysis tables are stored under the
base directory in a subdirectory named for the test.  That is, base/test.
If the test is not named, thes files are stored in base/output
Also, add a new option called "join" which joins the analysis files from several tests
into a single table used for plotting.