Clean up of the initialization of the base directory and shear.config.
We assume that the shear.config file is written only once, either manually,
or by this program.  Once it is written to the base dir, it is definitive.

Change code so it writes shear.config only once to the base directory.
Remove code in each of the subtasks which make modifications to the config.
shear.config is now read only once and not modified. It is also saved only once.