Code to modify footprint size and psf size and stamp size for various bias tests
The psf size is reduced to config.psfSize (psf_size in shear.config) if that is < the psf image.

The footprint size is reduced to config.footprintSize if that config item is specified.
    If it is not specified, code added to call SourceDetectionTask to find the footprint.  If no footprint
    is found, the galaxy is not processed.  If more than one is found, at attempt is made to find
    the footprint closest to the centroid, but within 2 arcsecs

The stamp size is the size of the exposure image.  For some measurement algorithms, this may bias
    the shear measurement, so we allow the stamp size to be reduced.