Modify analysis technique to treat e1 and e2 separately
The code used to attempt save the deviation (Dev) of e1 and e2 from g1 and g2.
I've changed the outputs so that it just save (g1,g2) and (e1Avg,e2Avg) for
each catalog.  This is all that is need since the plot program attempt to do a
linear fit between the measured average ellipticity and the applied shear.

Removed the "subfield catalog" which was used for debugging.

Added some comments, cleaned up whitespace and formatting.

Removed debugging code having to do with theta, the rotation angle

Stopped using the butler to get src.fits files