Modify the labelling to accomodate tests which have string names, like aDoubleGaussian.
Only the bottom panel will have a set of ticklabels, derived from the test name if
it is a string, or from the parameter value if one of the parameters:
nStampSize, nInitialRadii, nGrowFootprint.  The list testnumbers is used to
sequentially number the tests so that they will appear equally spaced on the
output plot if the labels are test names.