Changes to allow the analysis step to be done with a resample.  Code to recalculate the error.
analyzeShearTest.py runAnal() now takes a resample argument.  When set to a value > 0, it will do a
random resampling with replacement of the measurement catalog for each subfield.  The source catalog
is then analyzed as before, and the result are output to anal_[testname_].fits.

The replacement analysis must first be done N number of time with the "resample" argument set to
a different number each time so that N different analysis files are stored.

The error_bootstrap in runShear.py will then calculate the standard deviation of the means for
the population of resampled catalogs, and output the ratio of the original error calculation and
the one calculated from the bootstrap.